refactor(sync): move one-off migrations into migrate.go (#83 item 3b)
Pure code move, no behavior change. Relocates the two backward-compat
migrations out of the core import file into a dedicated migrate.go.

Moved from database.go -> internal/sync/migrate.go:
- removeLegacySQLite       (deletes pre-v0.3 _notion_sync.sqlite/.db)
- migrateToUUIDFilenames   (renames title-based .md to UUID filenames)

Both are unexported, same-package; all callers stay in sync (refresh.go,
database_test.go). migrateToUUIDFilenames still references localFileInfo
(unchanged, in database.go) — same-package cross-file ref, no signature
or import change.

database.go + migrate.go are committed together: the move is atomic, so
neither file compiles without the other.

Item 3a (extract agents.go to its own package) is intentionally NOT done
here — it requires breaking a sync<->agentsdoc import cycle (agents.go
reads the Version global; sync calls WriteAgentsMD), which is a design
change, not a move. Deferred to #87.

Verified: removed lines from database.go are byte-identical to migrate.go's
bodies; go build, go vet, go test ./... all pass.

Related to #83
